moved a few files and created folders
