+++ app/__init__.py
@@ -1,6 +1,6 @@
 #__init__ makes app a python package
 from flask import Flask
 
-app = Flask(__name__, template_folder='../templates')
+app = Flask(__name__)
 
 from app import routes
